requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-bootstrap3==11.0.0
django-oauth-toolkit==1.2.0
django-tables2==2.0.4
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
whitenoise==4.0
